NO MRG, TST: Attempt to build `master`
Have been running into some issues getting this package to build on
Windows in some cases. As the last build passed on `master` and recent
re-rendering build failed despite relatively little difference between
the two, this commit exists to simply test to see if `master` itself
will still build. If not, this implies that something else in the build
chain has changed, which is causing this issue. Hopefully comparing this
build and the last successful build on `master` will make this obvious
what is causing the problem.